chore(refactor): copy schema validation from render-markdown
## Details

Copies over the `render-markdown.nvim` schema validation implementation,
minus the bits that are not currently needed here.

Easier if I share implementations for things like this, at least to a
close approximation, in case any bugs or improvements are done on either
side. Logic is also much more concise and easy to reason about than
before.